[logger] Change all log_this decorators to new abstract one
