+++ superset/models/core.py
@@ -19,13 +19,12 @@
 from contextlib import closing
 from copy import copy, deepcopy
 from datetime import datetime
-import functools
 import json
 import logging
 import textwrap
 from typing import List
 
-from flask import current_app, escape, g, Markup, request
+from flask import escape, g, Markup, request
 from flask_appbuilder import Model
 from flask_appbuilder.models.decorators import renders
 from flask_appbuilder.security.sqla.models import User
@@ -54,7 +53,6 @@ from superset.models.user_attributes import UserAttribute
 from superset.utils import (
     cache as cache_util,
     core as utils,
-    log,
 )
 from superset.viz import viz_types
 from urllib import parse  # noqa
@@ -1128,58 +1126,6 @@ class Log(Model):
     dttm = Column(DateTime, default=datetime.utcnow)
     duration_ms = Column(Integer)
     referrer = Column(String(1024))
-
-    @classmethod
-    def log_this(cls, f):
-        """Decorator to log user actions"""
-        @functools.wraps(f)
-        def wrapper(*args, **kwargs):
-            user_id = None
-            if g.user:
-                user_id = g.user.get_id()
-            d = request.form.to_dict() or {}
-
-            # request parameters can overwrite post body
-            request_params = request.args.to_dict()
-            d.update(request_params)
-            d.update(kwargs)
-
-            slice_id = d.get('slice_id')
-            dashboard_id = d.get('dashboard_id')
-
-            try:
-                slice_id = int(
-                    slice_id or json.loads(d.get('form_data')).get('slice_id'))
-            except (ValueError, TypeError):
-                slice_id = 0
-
-            stats_logger.incr(f.__name__)
-            start_dttm = datetime.now()
-            value = f(*args, **kwargs)
-            duration_ms = (datetime.now() - start_dttm).total_seconds() * 1000
-
-            # bulk insert
-            try:
-                explode_by = d.get('explode')
-                records = json.loads(d.get(explode_by))
-            except Exception:
-                records = [d]
-
-            referrer = request.referrer[:1000] if request.referrer else None
-
-            action_logger = current_app.config.get('EVENT_LOGGER', log.DBEventLogger)()
-            action_logger.log(
-                user_id,
-                f.__name__,
-                records=records,
-                dashboard_id=dashboard_id,
-                slice_id=slice_id,
-                duration_ms=duration_ms,
-                referrer=referrer,
-            )
-            return value
-
-        return wrapper
 
 
 class FavStar(Model):
