+++ superset/views/core.py
@@ -43,7 +43,7 @@ from werkzeug.routing import BaseConverter
 from werkzeug.utils import secure_filename
 
 from superset import (
-    app, appbuilder, cache, conf, db, get_feature_flags, results_backend,
+    app, appbuilder, cache, conf, db, event_logger, get_feature_flags, results_backend,
     security_manager, sql_lab, viz)
 from superset.connectors.connector_registry import ConnectorRegistry
 from superset.connectors.sqla.models import AnnotationDatasource, SqlaTable
@@ -72,7 +72,6 @@ from .utils import bootstrap_user_data, get_datasource_info, get_form_data, get_
 config = app.config
 CACHE_DEFAULT_TIMEOUT = config.get('CACHE_DEFAULT_TIMEOUT', 0)
 stats_logger = config.get('STATS_LOGGER')
-log_this = models.Log.log_this
 DAR = models.DatasourceAccessRequest
 QueryStatus = utils.QueryStatus
 
@@ -680,7 +679,7 @@ class DashboardModelView(SupersetModelView, DeleteMixin):  # noqa
         return redirect(
             '/dashboard/export_dashboards_form?{}'.format(ids[1:]))
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/export_dashboards_form')
     def download_dashboards(self):
@@ -781,7 +780,7 @@ class KV(BaseSupersetView):
 
     """Used for storing and retrieving key value pairs"""
 
-    @log_this
+    @event_logger.log_this
     @has_access_api
     @expose('/store/', methods=['POST'])
     def store(self):
@@ -796,7 +795,7 @@ class KV(BaseSupersetView):
             json.dumps({'id': obj.id}),
             status=200)
 
-    @log_this
+    @event_logger.log_this
     @has_access_api
     @expose('/<key_id>/', methods=['GET'])
     def get_value(self, key_id):
@@ -815,7 +814,7 @@ class R(BaseSupersetView):
 
     """used for short urls"""
 
-    @log_this
+    @event_logger.log_this
     @expose('/<url_id>')
     def index(self, url_id):
         url = db.session.query(models.Url).filter_by(id=url_id).first()
@@ -830,7 +829,7 @@ class R(BaseSupersetView):
             flash('URL to nowhere...', 'danger')
             return redirect('/')
 
-    @log_this
+    @event_logger.log_this
     @has_access_api
     @expose('/shortner/', methods=['POST'])
     def shortner(self):
@@ -910,7 +909,7 @@ class Superset(BaseSupersetView):
             'requested': list(db_ds_names),
         }, status=201)
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/request_access/')
     def request_access(self):
@@ -958,7 +957,7 @@ class Superset(BaseSupersetView):
             datasource_names=', '.join([o.name for o in datasources]),
         )
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/approve')
     def approve(self):
@@ -1187,7 +1186,7 @@ class Superset(BaseSupersetView):
         payload = viz_obj.get_payload()
         return data_payload_response(*viz_obj.payload_json_and_has_error(payload))
 
-    @log_this
+    @event_logger.log_this
     @api
     @has_access_api
     @expose('/slice_json/<slice_id>')
@@ -1204,7 +1203,7 @@ class Superset(BaseSupersetView):
         )
         return self.generate_json(viz_obj)
 
-    @log_this
+    @event_logger.log_this
     @api
     @has_access_api
     @expose('/annotation_json/<layer_id>')
@@ -1223,7 +1222,7 @@ class Superset(BaseSupersetView):
         payload = viz_obj.get_payload()
         return data_payload_response(*viz_obj.payload_json_and_has_error(payload))
 
-    @log_this
+    @event_logger.log_this
     @api
     @has_access_api
     @handle_api_exception
@@ -1264,7 +1263,7 @@ class Superset(BaseSupersetView):
             samples=samples,
         )
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/import_dashboards', methods=['GET', 'POST'])
     def import_dashboards(self):
@@ -1275,7 +1274,7 @@ class Superset(BaseSupersetView):
             return redirect('/dashboard/list/')
         return self.render_template('superset/import_dashboards.html')
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/explorev2/<datasource_type>/<datasource_id>/')
     def explorev2(self, datasource_type, datasource_id):
@@ -1286,7 +1285,7 @@ class Superset(BaseSupersetView):
             datasource_id=datasource_id,
             **request.args))
 
-    @log_this
+    @event_logger.log_this
     @has_access
     @expose('/explore/<datasource_type>/<datasource_id>/', methods=['GET', 'POST'])
     @expose('/explore/', methods=['GET', 'POST'])
@@ -2215,7 +2214,7 @@ class Superset(BaseSupersetView):
         edit_mode = request.args.get('edit') == 'true'
 
         # Hack to log the dashboard_id properly, even when getting a slug
-        @log_this
+        @event_logger.log_this
         def dashboard(**kwargs):  # noqa
             pass
         dashboard(
@@ -2254,14 +2253,14 @@ class Superset(BaseSupersetView):
         )
 
     @api
-    @log_this
+    @event_logger.log_this
     @expose('/log/', methods=['POST'])
     def log(self):
         return Response(status=200)
 
     @has_access
     @expose('/sync_druid/', methods=['POST'])
-    @log_this
+    @event_logger.log_this
     def sync_druid_source(self):
         """Syncs the druid datasource in main db with the provided config.
 
@@ -2313,7 +2312,7 @@ class Superset(BaseSupersetView):
 
     @has_access
     @expose('/sqllab_viz/', methods=['POST'])
-    @log_this
+    @event_logger.log_this
     def sqllab_viz(self):
         SqlaTable = ConnectorRegistry.sources['table']
         data = json.loads(request.form.get('data'))
@@ -2358,7 +2357,7 @@ class Superset(BaseSupersetView):
 
     @has_access
     @expose('/table/<database_id>/<table_name>/<schema>/')
-    @log_this
+    @event_logger.log_this
     def table(self, database_id, table_name, schema):
         schema = utils.parse_js_uri_path_item(schema, eval_undefined=True)
         table_name = utils.parse_js_uri_path_item(table_name)
@@ -2418,7 +2417,7 @@ class Superset(BaseSupersetView):
 
     @has_access
     @expose('/extra_table_metadata/<database_id>/<table_name>/<schema>/')
-    @log_this
+    @event_logger.log_this
     def extra_table_metadata(self, database_id, table_name, schema):
         schema = utils.parse_js_uri_path_item(schema, eval_undefined=True)
         table_name = utils.parse_js_uri_path_item(table_name)
@@ -2430,7 +2429,7 @@ class Superset(BaseSupersetView):
     @has_access
     @expose('/select_star/<database_id>/<table_name>')
     @expose('/select_star/<database_id>/<table_name>/<schema>')
-    @log_this
+    @event_logger.log_this
     def select_star(self, database_id, table_name, schema=None):
         mydb = db.session.query(
             models.Database).filter_by(id=database_id).first()
@@ -2451,7 +2450,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/cached_key/<key>/')
-    @log_this
+    @event_logger.log_this
     def cached_key(self, key):
         """Returns a key from the cache"""
         resp = cache.get(key)
@@ -2461,7 +2460,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/cache_key_exist/<key>/')
-    @log_this
+    @event_logger.log_this
     def cache_key_exist(self, key):
         """Returns if a key from cache exist"""
         key_exist = True if cache.get(key) else False
@@ -2471,7 +2470,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/results/<key>/')
-    @log_this
+    @event_logger.log_this
     def results(self, key):
         """Serves a key off of the results backend"""
         if not results_backend:
@@ -2512,7 +2511,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/stop_query/', methods=['POST'])
-    @log_this
+    @event_logger.log_this
     def stop_query(self):
         client_id = request.form.get('client_id')
         try:
@@ -2528,7 +2527,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/validate_sql_json/', methods=['POST', 'GET'])
-    @log_this
+    @event_logger.log_this
     def validate_sql_json(self):
         """Validates that arbitrary sql is acceptable for the given database.
         Returns a list of error/warning annotations as json.
@@ -2593,7 +2592,7 @@ class Superset(BaseSupersetView):
 
     @has_access_api
     @expose('/sql_json/', methods=['POST', 'GET'])
-    @log_this
+    @event_logger.log_this
     def sql_json(self):
         """Runs arbitrary sql and returns and json"""
         async_ = request.form.get('runAsync') == 'true'
@@ -2725,7 +2724,7 @@ class Superset(BaseSupersetView):
 
     @has_access
     @expose('/csv/<client_id>')
-    @log_this
+    @event_logger.log_this
     def csv(self, client_id):
         """Download the query results as csv."""
         logging.info('Exporting CSV file [{}]'.format(client_id))
@@ -2770,7 +2769,7 @@ class Superset(BaseSupersetView):
     @handle_api_exception
     @has_access
     @expose('/fetch_datasource_metadata')
-    @log_this
+    @event_logger.log_this
     def fetch_datasource_metadata(self):
         datasource_id, datasource_type = (
             request.args.get('datasourceKey').split('__'))
@@ -2813,7 +2812,7 @@ class Superset(BaseSupersetView):
 
     @has_access
     @expose('/search_queries')
-    @log_this
+    @event_logger.log_this
     def search_queries(self) -> Response:
         """
         Search for previously run sqllab queries. Used for Sqllab Query Search
